# docs/swagger.yaml
basePath: /api
definitions:
  api.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  domain.Account:
    properties:
      balance:
        description: stored as minor units (cents)
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      id:
        type: string
      ownerID:
        type: string
      status:
        $ref: '#/definitions/domain.AccountStatus'
      updatedAt:
        type: string
      version:
        type: integer
    type: object
  domain.AccountStatus:
    enum:
    - ACTIVE
    - FROZEN
    - CLOSED
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusFrozen
    - StatusClosed
  domain.EventType:
    enum:
    - ACCOUNT_CREATED
    - ACCOUNT_CREDITED
    - ACCOUNT_DEBITED
    - ACCOUNT_STATUS_CHANGED
    - TRANSFER_INITIATED
    - TRANSFER_COMPLETED
    - TRANSFER_REVERSED
    type: string
    x-enum-varnames:
    - EventAccountCreated
    - EventAccountCredited
    - EventAccountDebited
    - EventAccountStatusChanged
    - EventTransferInitiated
    - EventTransferCompleted
    - EventTransferReversed
  domain.LedgerEvent:
    properties:
      aggregateID:
        type: string
      createdAt:
        type: string
      eventType:
        $ref: '#/definitions/domain.EventType'
      id:
        type: string
      payload:
        items:
          type: integer
        type: array
      version:
        type: integer
    type: object
  services.ChangeStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/domain.AccountStatus'
        enum:
        - ACTIVE
        - FROZEN
        - CLOSED
    required:
    - status
    type: object
  services.CreateAccountRequest:
    properties:
      currency:
        type: string
      ownerId:
        type: string
    required:
    - currency
    - ownerId
    type: object
  services.CreditRequest:
    properties:
      amount:
        type: integer
      currency:
        type: string
      reference:
        type: string
    required:
    - amount
    - currency
    - reference
    type: object
  services.DebitRequest:
    properties:
      amount:
        type: integer
      currency:
        type: string
      reference:
        type: string
    required:
    - amount
    - currency
    - reference
    type: object
  services.TransferRequest:
    properties:
      amount:
        type: integer
      currency:
        type: string
      sourceAccountId:
        type: string
      targetAccountId:
        type: string
    required:
    - amount
    - currency
    - sourceAccountId
    - targetAccountId
    type: object
host: localhost:8080
info:
  contact:
    email: ledger@example.com
    name: Ledger Team
  description: 'Write-side of the CQRS ledger: account creation, credits, debits,
    transfers.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Ledger Command Service API
  version: "1.0"
paths:
  /v1/accounts:
    post:
      consumes:
      - application/json
      description: Creates a ledger account for an owner
      parameters:
      - description: Create account request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create a new account
      tags:
      - accounts
  /v1/accounts/{accountId}/credits:
    post:
      consumes:
      - application/json
      description: Adds funds to an account
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Idempotency key (UUID)
        in: header
        name: Idempotency-Key
        type: string
      - description: Credit request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CreditRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.LedgerEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Credit an account
      tags:
      - accounts
  /v1/accounts/{accountId}/debits:
    post:
      consumes:
      - application/json
      description: Withdraws funds from an account
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Idempotency key (UUID)
        in: header
        name: Idempotency-Key
        type: string
      - description: Debit request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.DebitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.LedgerEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Debit an account
      tags:
      - accounts
  /v1/accounts/{accountId}/events:
    get:
      description: Returns the full audit trail of events for an account
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LedgerEvent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get raw event log
      tags:
      - accounts
  /v1/accounts/{accountId}/status:
    patch:
      consumes:
      - application/json
      description: Sets an account to ACTIVE, FROZEN, or CLOSED
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Status change request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.ChangeStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LedgerEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Change account status
      tags:
      - accounts
  /v1/transfers:
    post:
      consumes:
      - application/json
      description: 'Initiates a choreography saga: debits source, credits target'
      parameters:
      - description: Idempotency key (UUID)
        in: header
        name: Idempotency-Key
        type: string
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.TransferRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/domain.LedgerEvent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Initiate a transfer
      tags:
      - transfers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"